Create fireball cannon and moved death logic to one script to avoid code duplication, using tags.
- Fireball cannon
- Avoid code duplication by using tags and death in one script
