+++ Assets/Prefab/Acid Trap.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 30cbf7d725456c047b66aae94a51dad5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
