+++ Assets/Prefab/Main Camera.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f5620d412a5551b418a4741c8e199c74
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
