+++ Assets/Prefab/Stone Platform.prefab
@@ -11,7 +11,7 @@ GameObject:
   - component: {fileID: 3206338880147713339}
   - component: {fileID: 1563612007423405215}
   - component: {fileID: 3816255976948711164}
-  m_Layer: 0
+  m_Layer: 3
   m_Name: Stone Platform
   m_TagString: Untagged
   m_Icon: {fileID: 0}
