+++ Assets/Script/DeathScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 751e92d174f96924d976947ea246f43b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
